Plotting of the NBSS - Normalized Biovolume Size Spectra - of a given sample

Plotting of the NBSS - Normalized Biovolume Size Spectra - of a given sample

Usage

NBSS.plot(planktotable, ESD = TRUE)

Arguments

ESD

If ESD is true, by default, the x axis is expressed as ESD (Eqivalent spherical diameter for a given volume) in µm. Otherwise, volumes will be on the x-axis in mm^3

objects

A data table containing objects as rows and details on these objects in the columns. This argument will be passed in the NBSS function

samples

A data table containing samples in rows and details about the latter in columns.

sample_name

A string containing the name of the sample from which to extract objects and do an NBSS analysis.

Value

an object of the type ggplot which can then be plotted.

Examples

p <- NBSS.plot(objects=objects,samples=samples,sample_name=sample_id,ESD=TRUE)
#> Error in NBSS.plot(objects = objects, samples = samples, sample_name = sample_id,     ESD = TRUE): unused arguments (objects = objects, samples = samples, sample_name = sample_id)
p 
#> Error in eval(expr, envir, enclos): object 'p' not found